bugfix: accessing floatValue is wrong when handing Double.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@23547 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/value/AbstractExpressionValueVisitor.java
@@ -603,8 +603,8 @@ public abstract class AbstractExpressionValueVisitor
         break;
       }
       case DOUBLE: {
-        double lVal = l.floatValue();
-        double rVal = r.floatValue();
+        double lVal = l.doubleValue();
+        double rVal = r.doubleValue();
 
         if (Double.isNaN(lVal) || Double.isNaN(rVal)) {
           return new NumericValue(op == BinaryOperator.NOT_EQUALS ? 1L : 0L);
